have an req/rep socket to control collator
This will let us use push/pull sockets on the collator for
the actual frames. Currently using req/rep from the client
thread to the collator thread is requiring multiple context
switches for every frame send (as the client is blocking on
the ack).

The overhead for this isn't a problem for databursts, but
this was bottlenecking throughput waiting on acks, and slaying
host CPUs with hundreds of thousands of context switches a
second.

In this case, the OOB req/rep socket doesn't do anything most
of the time. We are now free to change the main data path from
the client to something that can actually queue multiple messages.