Update build_spool_path() test to deal with all the changes
The test's assumptions are now fixed (no more double-slashes in the
path) and rewritten in large parts to be sane. Why were we doing length
checks? Why were we rolling our own char-by-char string comparison?
strcmp can do that for us just fine, doubly so when we have a guaranteed
length boundary to pass to strncmp.