Add a readback test for source dicts
I've noted that this code could be a *lot* shorter if I didn't care
about leaking memory and FDs, but I feel like I'd be trolling myself if
I did that.

Also, yes, I'm probably not making effective use of Glib's testing
functionality, but it looked like the overhead in line of code wouldn't
really be worth it to handle setup and teardown at this scale.