Fix marquise_ctx structure to avoid use of arrays
This is necessary due to Python CFFI giving us grief (it's how we made
the pymarquise binding. It appears to be miscalculating the expected
size of the struct, causing verification errors. As best I can tell it
sees the char* and misses that it's an array of char* pointers.

It's not quite ideal, but this simplistic change will do the job, making
each pointer explicit.