Set the LSB of all sourcedict addresses to zero
We've identified a problem with the way that sourcedicts are stored in
the vault. Normally, we generate an address and pass it along with
points and sourcedicts. For simple points, the address gets coerced to
an even number by setting the least significant bit to zero. Extended
points have their LSB set to one.

The same is not happening for sourcedicts. This means that it's possible
for data points and their associated sourcedict to be stored at
different addresses. This is bad, because we can enumerate the
sourcedicts, but won't necessarily find data points at the corresponding
address.

We've determined that the correct behaviour, at least if we follow the
"official" Marquise client, is to set the LSB of all sourcedicts to
zero.